Add batch mode and Dockerfile for running as a container
This change adds getopt support for changing various command line
args, and also adds a new "batch" mode which runs everything and
spits it out in base64 to stdout.  This can be used in a container
(with included Dockerfile), although with the caveat that it can't
restart the Docker daemon.

I've been running it with the command:

$ docker run -it --rm -v /boot:/boot -v /var/run/docker.sock:/var/run/docker.sock -v /var/lib/docker:/var/lib/docker -v /var/log:/var/log -v /etc/sysconfig:/etc/sysconfig -v /etc/default:/etc/default pdevine/dsinfo:latest > foo.base64

Signed-off-by: Patrick Devine <patrick.devine@docker.com>